+++ host/esutil/client_v7.go
@@ -18,34 +18,34 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package esutils
+package esutil
 
 import (
 	"fmt"
 	"io/ioutil"
 	"time"
 
-	"github.com/olivere/elastic"
+	"github.com/olivere/elastic/v7"
 	"github.com/stretchr/testify/suite"
 )
 
 type (
-	v6Client struct {
+	v7Client struct {
 		client *elastic.Client
 	}
 )
 
-func newV6Client(url string) (*v6Client, error) {
+func newV7Client(url string) (*v7Client, error) {
 	esClient, err := elastic.NewClient(
 		elastic.SetURL(url),
 		elastic.SetRetrier(elastic.NewBackoffRetrier(elastic.NewExponentialBackoff(128*time.Millisecond, 513*time.Millisecond))),
 	)
-	return &v6Client{
+	return &v7Client{
 		client: esClient,
 	}, err
 }
 
-func (es *v6Client) PutIndexTemplate(s suite.Suite, templateConfigFile, templateName string) {
+func (es *v7Client) PutIndexTemplate(s suite.Suite, templateConfigFile, templateName string) {
 	// This function is used exclusively in tests. Excluding it from security checks.
 	// #nosec
 	template, err := ioutil.ReadFile(templateConfigFile)
@@ -55,7 +55,7 @@ func (es *v6Client) PutIndexTemplate(s suite.Suite, templateConfigFile, template
 	s.Require().True(putTemplate.Acknowledged)
 }
 
-func (es *v6Client) CreateIndex(s suite.Suite, indexName string) {
+func (es *v7Client) CreateIndex(s suite.Suite, indexName string) {
 	exists, err := es.client.IndexExists(indexName).Do(createContext())
 	s.Require().NoError(err)
 	if exists {
@@ -69,20 +69,20 @@ func (es *v6Client) CreateIndex(s suite.Suite, indexName string) {
 	s.Require().True(createTestIndex.Acknowledged)
 }
 
-func (es *v6Client) DeleteIndex(s suite.Suite, indexName string) {
+func (es *v7Client) DeleteIndex(s suite.Suite, indexName string) {
 	deleteTestIndex, err := es.client.DeleteIndex(indexName).Do(createContext())
 	s.Nil(err)
 	s.True(deleteTestIndex.Acknowledged)
 }
 
-func (es *v6Client) PutMaxResultWindow(indexName string, maxResultWindow int) error {
+func (es *v7Client) PutMaxResultWindow(indexName string, maxResultWindow int) error {
 	_, err := es.client.IndexPutSettings(indexName).
 		BodyString(fmt.Sprintf(`{"max_result_window" : %d}`, maxResultWindow)).
 		Do(createContext())
 	return err
 }
 
-func (es *v6Client) GetMaxResultWindow(indexName string) (string, error) {
+func (es *v7Client) GetMaxResultWindow(indexName string) (string, error) {
 	settings, err := es.client.IndexGetSettings(indexName).Do(createContext())
 	if err != nil {
 		return "", err
